Add typecast for param->time_to_die comp

+++ philo/src/parse_args.c
@@ -6,7 +6,7 @@
 /*   By: stena-he <stena-he@student.42wolfsburg.    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/04 01:54:01 by stena-he          #+#    #+#             */
-/*   Updated: 2023/01/10 22:08:07 by stena-he         ###   ########.fr       */
+/*   Updated: 2023/01/10 23:34:07 by stena-he         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -73,7 +73,7 @@ int	parse_args(int argc, char **argv, t_param *param)
 		param->times_must_eat = 0;
 	if (param->n_philo < 0)
 		return (-1);
-	if (param->time_to_die < 0)
+	if ((int)param->time_to_die < 0)
 		return (-1);
 	if (param->time_to_eat < 0)
 		return (-1);
